config: reject extended sources in photocal
The reference catalog may not have as good star/galaxy discrimination
as our data (e.g., PS1 PV1.2), so reject sources that we consider to be
extended.

Note that this treats classification.extendedness as a flag which it is
not, but since we are treating it as a boolean value in floating-point
(0.0 or 1.0), this works fine.