HscMapper: remove old filter names (W-S-*) as independent filters
The W-S-* filter names were used on Suprime-Cam, and we expected
they might be used on HSC, but they were not.  They should not be
included as individual filters in HscMapper.filters because they
do not exist.  They are still listed as aliases in case that is
ever required; this is no burden (does not use integers required
for object IDs in coadds).