genDefectRegistry: support multiple versions of same CCD defects
Multiple versions of the same CCD defects were thrown in the
database with the same end date (way in the future), so that
the query to find the defect ended up with multiple rows.
Instead, have one defect valid until the next, and only the
last is valid into the future.