Set BAD mask for dead amps instead of SAT.
Original saturationDetection() is using afwDetection to identify saturated pixels
but this method is slow in the case whole amp region is saturated.
So replace it with a simple numpy.where() which works well.
If entire region is set to SAT, saturationInterpolation() will try to interpolate
whole amplifier and it takes time again.
So BAD mask will be set to dead amp instead of SAT.
This change require crosstalk code to handle BAD mask.