isr: add brighter fatter correction as calibration option.
This implements the brighter fatter correction method of
Will Coulton.  He generated a kernel from flat field images that
can be applied iteratively to an image to correct for intensity
dependent PSFs.  We still need to investigate when to stop iterating,
but the current settings seem to work.  This option is currently
disabled until it has been further tested.

This change does not include the code to generate the kernel.