config: disable 2D bias subtraction in general processing
Overscan subtraction removes most of the bias, and though it leaves
a line running down the center of the CCD, this is usually below the
noise level.

Explicitly did not disable bias subtraction for dark construction
because we would convert that center line to a count rate, which
(depending on the exposure times involved) could cause it to
become visible.