genCalibRegistry: fix overlap between validStart,validEnd
When registering calibrations, genCalibRegistry.py adjusts the
validStart and validEnd dates in an attempt to avoid conflicts.
However, it sets one's validStart date to another's validEnd date,
so that an exposure on that date would have two matches. Tweaking
the validEnd by a day removes the overlap.

This bug was first noted by Hsin-Fang Chiang when porting the
functionality to LSST (DM-3766).