config: reduce coadd subregionSize
The HSC SSP Ultra-Deep currently has a total of 85 exposures in y-band,
so that reading 2k square blocks exceeds 4 GB/core. Reducing the block
size reduces the memory use at the expense of a bit of read inefficiency.
This choice (200 rows, for a 4k patch width) should give us plenty of
room for growth.

To attempt to gain back some read efficiency, use full rows (the 10k
width set here is much larger than the typical patch size).

Implemented as an obs_subaru override instead of changing the default in
pipe_tasks because the default is fine for most other applications.