config: reduce background mesh size for detection on coadds
We have determined that we want to use a binSize=128 pixels
for the background mesh in coadd detection. This will reduce
the load on the deblender, and optimise the quality of the
detection and photometry for small (almost-)unresolved sources.
Those interested in science on large sources can restore the
background and make their own specialist measurements.

This change was provided by Hiroyuki Ikeda.